+++ src/django_safe_template_engine/trusted_tags.py
@@ -34,6 +34,12 @@ register.tag(ifchanged)
 register.tag(lorem)
 register.tag(now)
 
+if VERSION >= (6, 0):
+    from django.template.defaulttags import partial_func, partialdef_func
+
+    register.tag("partial", partial_func)
+    register.tag("partialdef", partialdef_func)
+
 if VERSION >= (5, 1):
     from django.template.defaulttags import querystring